Move benchmarks into cargo benches directory
